% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JBQ.R
\name{JBQ-class}
\alias{JBQ-class}
\alias{JBQ}
\title{The JBQ | JAX BioConnect Query tool}
\description{
The JBQ | JAX BioConnect Query tool

The JBQ | JAX BioConnect Query tool
}
\details{
Class \code{JBQ} wraps a JBQ functions
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path_to_file_or_dir}}{character, File path to a .zip file or an uncompressed directory}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-JBQ-new}{\code{JBQ$new()}}
\item \href{#method-JBQ-show_studies}{\code{JBQ$show_studies()}}
\item \href{#method-JBQ-show_files}{\code{JBQ$show_files()}}
\item \href{#method-JBQ-get_assay_samples}{\code{JBQ$get_assay_samples()}}
\item \href{#method-JBQ-get_assay_sources_samples}{\code{JBQ$get_assay_sources_samples()}}
\item \href{#method-JBQ-get_study_sources}{\code{JBQ$get_study_sources()}}
\item \href{#method-JBQ-get_files}{\code{JBQ$get_files()}}
\item \href{#method-JBQ-jbq_version}{\code{JBQ$jbq_version()}}
\item \href{#method-JBQ-read_package}{\code{JBQ$read_package()}}
\item \href{#method-JBQ-jbq_command}{\code{JBQ$jbq_command()}}
\item \href{#method-JBQ-parse}{\code{JBQ$parse()}}
\item \href{#method-JBQ-test}{\code{JBQ$test()}}
\item \href{#method-JBQ-get_version}{\code{JBQ$get_version()}}
\item \href{#method-JBQ-finalize}{\code{JBQ$finalize()}}
\item \href{#method-JBQ-print}{\code{JBQ$print()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JBQ-new"></a>}}
\if{latex}{\out{\hypertarget{method-JBQ-new}{}}}
\subsection{Method \code{new()}}{
private fields


Initialization method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JBQ$new(path_to_file_or_dir = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path_to_file_or_dir}}{character. File path to a .zip file or an uncompressed directory}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JBQ-show_studies"></a>}}
\if{latex}{\out{\hypertarget{method-JBQ-show_studies}{}}}
\subsection{Method \code{show_studies()}}{
show_studies, show studies
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JBQ$show_studies()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
character, output
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JBQ-show_files"></a>}}
\if{latex}{\out{\hypertarget{method-JBQ-show_files}{}}}
\subsection{Method \code{show_files()}}{
show_files, show files
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JBQ$show_files()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
character, output
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JBQ-get_assay_samples"></a>}}
\if{latex}{\out{\hypertarget{method-JBQ-get_assay_samples}{}}}
\subsection{Method \code{get_assay_samples()}}{
get_assay_samples, get assay samples
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JBQ$get_assay_samples()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
dataframe
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JBQ-get_assay_sources_samples"></a>}}
\if{latex}{\out{\hypertarget{method-JBQ-get_assay_sources_samples}{}}}
\subsection{Method \code{get_assay_sources_samples()}}{
get_assay_sources_samples, get assay sources samples
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JBQ$get_assay_sources_samples()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
dataframe
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JBQ-get_study_sources"></a>}}
\if{latex}{\out{\hypertarget{method-JBQ-get_study_sources}{}}}
\subsection{Method \code{get_study_sources()}}{
get_study_sources, get study sources
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JBQ$get_study_sources()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
dataframe
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JBQ-get_files"></a>}}
\if{latex}{\out{\hypertarget{method-JBQ-get_files}{}}}
\subsection{Method \code{get_files()}}{
get_files, download files
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JBQ$get_files(file_name_filter)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file_name_filter}}{character. file name filter, "all" for all file}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
character, output
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JBQ-jbq_version"></a>}}
\if{latex}{\out{\hypertarget{method-JBQ-jbq_version}{}}}
\subsection{Method \code{jbq_version()}}{
get_jbq_version, get jbq version
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JBQ$jbq_version()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
dataframe
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JBQ-read_package"></a>}}
\if{latex}{\out{\hypertarget{method-JBQ-read_package}{}}}
\subsection{Method \code{read_package()}}{
read, do a system call with a operation and return the call output
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JBQ$read_package(optional = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{optional}}{character. JBQ optional}

\item{\code{command}}{character. JBQ command}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
character vector, the output of the system JBQ call
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JBQ-jbq_command"></a>}}
\if{latex}{\out{\hypertarget{method-JBQ-jbq_command}{}}}
\subsection{Method \code{jbq_command()}}{
system_call, do a system call with a operation and return the call output
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JBQ$jbq_command(command = NULL, operation = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{command}}{character. JBQ command}

\item{\code{operation}}{character. JBQ operation}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
character vector, the output of the system JBQ call
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JBQ-parse"></a>}}
\if{latex}{\out{\hypertarget{method-JBQ-parse}{}}}
\subsection{Method \code{parse()}}{
parse, parse the system call output to dataframe
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JBQ$parse(output, data_tag)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{output}}{character. system call output}

\item{\code{data_tag}}{character. data tag}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
dataframe, parsed the dataframe from the output
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JBQ-test"></a>}}
\if{latex}{\out{\hypertarget{method-JBQ-test}{}}}
\subsection{Method \code{test()}}{
test method.
This is a test function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JBQ$test()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JBQ-get_version"></a>}}
\if{latex}{\out{\hypertarget{method-JBQ-get_version}{}}}
\subsection{Method \code{get_version()}}{
get_version, get current version
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JBQ$get_version()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
character, version
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JBQ-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-JBQ-finalize}{}}}
\subsection{Method \code{finalize()}}{
finalize
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JBQ$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JBQ-print"></a>}}
\if{latex}{\out{\hypertarget{method-JBQ-print}{}}}
\subsection{Method \code{print()}}{
print
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JBQ$print()}\if{html}{\out{</div>}}
}

}
}
